Got rid of ticket information in URL
Use a temporary cookie to store the ticket information after fetching
one from the CAS server and redirect to the root. If a ticket cookie
exists on a request to the root, use that cookie to authenticate. If
it works, then we clear the ticket cookie and assign a new cookie
containing the netid. Otherwise, we redirect to the CAS server.